Create FUNDING.yml
## 💖 Support Orbital Screen
At the top of every .py, .js, .ts, or .sh file, include this header:

plaintext
Orbital Screen — Modular Cockpit Integration System  
Copyright (C) 2025 Orbital

This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY, without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
You should have received a copy of the GNU General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.  

💖 Support Orbital Screen

Orbital Screen is built for creators, sustained by visionaries.  
If you believe in open orchestration and creative sovereignty, consider sponsoring the project.

[Sponsor this project](https://buymeacoffee.com/orbital)  
[More ways to support](https://orbital-screen.io/sponsor)



Signed-off-by: topcat042000-maker <topcat042000@gmail.com>